// src/app/@core/model/rest/cards/prepaid-cards/request-new-owner-data-res.model.ts
export interface PrepaidRequestCardOwnerDataResModel {
    birthDate: string;
    city: string;
    companyEmbossingName: any;
    firstName: string;
    gender: string;
    generatedEmbossingNames: boolean;
    lastName: string;
    mobileNumber: string;
    nationalId: string;
    nationality: string;
    ownerEmbossingName: string[];
}